pyproject.toml
[project]
name = "hai-sensemaking"
version = "0.1.0"
description = "Human-AI Sensemaking Canvas"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.39.0",
    "openai>=2.26.0",
    "python-dotenv>=1.2.0",
    "PyYAML>=6.0.0",
]

[project.scripts]
app = "backend.main:app"
